Replaced int with float

+++ 3.py
@@ -1,3 +1,3 @@
-R = int(input("Enter a radius of the circle "))
+R = float(input("Enter a radius of the circle "))
 area = 3.14 * R ** 2 
 print("The area of the circle is " + str(area))
